refactor: fix naming problems
